+++ example/_shared/lib/todo/view/web/todo_input_view.dart
@@ -1,31 +1,41 @@
-import 'dart:html';
-
+import 'dart:js_interop';
+import 'package:web/web.dart';
 import 'package:wire/wire.dart';
 import 'package:wire_example_shared/todo/const/view_signals.dart';
 import 'package:wire_example_shared/todo/data/dto/input_dto.dart';
+import 'package:wire_example_shared/todo/data/vo/todo_vo.dart';
 import 'package:wire_example_shared/todo/view/web/base/dom_element_view.dart';
 
 class TodoInputView extends DomElement {
-  TodoInputView(InputElement dom) : super(dom) {
+  TodoInputView(super.dom) {
     Wire.add(this, ViewSignals.CLEAR_INPUT, _clearInput);
-
+    print('> TodoInputView -> setup');
     dom
-      ..text = ''
-      ..onKeyPress.listen((e) async {
-        if (e.keyCode == KeyCode.ENTER) {
-          _onInputEnter();
-        }
-      });
+      ..textContent = ''
+      ..addEventListener(
+        'keyup',
+        (KeyboardEvent e) {
+          print('> TodoInputView -> keyup: ${e}');
+          if (e.keyCode == KeyCode.ENTER) {
+            _onInputEnter();
+          }
+        }.toJS,
+      );
   }
 
-  InputElement get input => dom as InputElement;
+  HTMLInputElement get input => dom as HTMLInputElement;
 
-  Future<void> _clearInput(data, wid) async {
+  Future<void> _clearInput(_, int __) async {
     input.value = '';
   }
 
   void _onInputEnter() {
-    final dto = InputDTO(input.value!, '');
-    Wire.send(ViewSignals.INPUT, payload: dto);
+    final dto = InputDTO(input.value, '');
+    Wire.send(ViewSignals.INPUT, payload: dto).then((WireSendResults results) {
+      print('> TodoInputView -> send results: ${results.list}');
+      if (results.list.isEmpty) return;
+      final result = results.list.singleWhere((item) => item is TodoVO) as TodoVO?;
+      if (result != null) print('> \tresult: ${result.text} - created');
+    });
   }
 }