+++ lib/mixin/mixin_with_wire_data.dart
@@ -1,4 +1,4 @@
-library wire;
+library;
 
 import 'package:wire/wire.dart';
 
@@ -7,25 +7,23 @@ mixin WireMixinWithWireData {
   bool has(String dataKey) => Wire.data(dataKey).isSet;
 
   Future<T> get<T>(String dataKey) => Future(() {
-        if (has(dataKey)) {
-          return Future.value(Wire.data(dataKey).value as T);
-        } else {
-          return Future.error('Error: missing data on key - ${dataKey}');
-        }
-      });
+    if (has(dataKey)) {
+      return Future.value(Wire.data<T>(dataKey).value);
+    } else {
+      return Future.error('Error: missing data on key - ${dataKey}');
+    }
+  });
 
-  Future<Map<String, dynamic>> getMany(List<String> many) =>
-      Future.wait(many.map((e) async => get(e))).then((List<dynamic> results, {int index = 0}) {
-        return many.fold<Map<String, dynamic>>(
-            {}, (Map<String, dynamic> p, key) => p..putIfAbsent(key, () => results[index++]));
-      });
+  Future<Map<String, dynamic>> getMany(List<String> many) => Future.wait(many.map((e) async => get(e))).then((List<dynamic> results, {int index = 0}) {
+    return many.fold<Map<String, dynamic>>({}, (Map<String, dynamic> p, key) => p..putIfAbsent(key, () => results[index++]));
+  });
 
-  void update(String dataKey, {dynamic data, bool refresh = true}) {
-    // log('> MixinWithWireData -> update(${dataKey}): $data');
+  void update<T>(String dataKey, {T? data, bool refresh = true}) {
+    print('> MixinWithWireData -> update(${dataKey}): $data - ${data.runtimeType}');
     if (data != null) {
-      Wire.data(dataKey, value: data);
+      Wire.data<T>(dataKey, value: data);
     } else if (refresh) {
-      Wire.data(dataKey).refresh();
+      Wire.data<T>(dataKey).refresh();
     }
   }